% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxy.R
\name{cmdiff}
\alias{cmdiff}
\title{Cmdiff colour palette}
\arguments{
\item{n}{integer giving the number of colours (>= 1) to be produced;
currently it does not make sense to ask for more than 256}
}
\value{
Vector of R colours (\code{'#RRGGBB'} strings) of length \code{n}.
}
\description{
The diff colormap is diverging, with one side shades of blues and one side shades of browns.
}
\examples{
z <- xtabs(weight~Time+Chick, ChickWeight)
x <- sort(as.numeric(rownames(z)))
y <- sort(as.numeric(colnames(z)))
image(
	x = x, y = y, z = z, col = cmdiff(100),
 xlab = 'Time', ylab = 'Chick'
)
}
\keyword{color}
