\name{cmocean-package}
\alias{cmocean-package}
\alias{cmocean}
\docType{package}
\title{
\packageTitle{cmocean}
}
\description{
\packageDescription{cmocean}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cmocean}
\packageIndices{cmocean}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{cmocean}

Maintainer: \packageMaintainer{cmocean}
}
\references{
Thyng, K. M., Greene, C. A., Hetland, R. D., Zimmerle, H. M., & DiMarco, S. F. (2016). True colors of oceanography. Oceanography, 29(3), 10.
}
\keyword{ package }
\keyword{ color }
\examples{
levelplot(weight ~ Time * Chick, ChickWeight, col.regions = thermal(100))
}
