% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxy.R
\name{delta}
\alias{delta}
\title{Delta colour palette}
\arguments{
\item{n}{integer giving the number of colours (>= 1) to be produced;
currently it does not make sense to ask for more than 256}
}
\value{
Vector of R colours (\code{'#RRGGBB'} strings) of length \code{n}.
}
\description{
The delta colormap is diverging from darker blues to just off-white through shades of yellow green and could be used to represent diverging velocity values around a critical value (usually zero). This colormap was inspired by Francesca Samsel’s similar colormap, but generated from scratch using the \sQuote{viscm} tool.
}
\examples{
z <- xtabs(weight~Time+Chick, ChickWeight)
x <- sort(as.numeric(rownames(z)))
y <- sort(as.numeric(colnames(z)))
image(
	x = x, y = y, z = z, col = delta(100),
 xlab = 'Time', ylab = 'Chick'
)
}
\keyword{color}
