% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxy.R
\name{haline}
\alias{haline}
\title{Haline colour palette}
\arguments{
\item{n}{integer giving the number of colours (>= 1) to be produced;
currently it does not make sense to ask for more than 256}
}
\value{
Vector of R colours (\code{'#RRGGBB'} strings) of length \code{n}.
}
\description{
The haline colormap is sequential, and might be used with dark blue representing lower salinity or fresher water, transitioning through greens to light yellow representing increased salinity or saltier water. This colormap is based on matplotlib’s \dQuote{YlGnBu}, but was recreated from scratch using the \sQuote{viscm} tool.
}
\examples{
z <- xtabs(weight~Time+Chick, ChickWeight)
x <- sort(as.numeric(rownames(z)))
y <- sort(as.numeric(colnames(z)))
image(
	x = x, y = y, z = z, col = haline(100),
 xlab = 'Time', ylab = 'Chick'
)
}
\keyword{color}
