% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxy.R
\name{ice}
\alias{ice}
\title{Ice colour palette}
\arguments{
\item{n}{integer giving the number of colours (>= 1) to be produced;
currently it does not make sense to ask for more than 256}
}
\value{
Vector of R colours (\code{'#RRGGBB'} strings) of length \code{n}.
}
\description{
The ice colormap is sequential from very dark blue (almost black) to very light blue (almost white). A use for this could be representations of sea ice.
}
\examples{
z <- xtabs(weight~Time+Chick, ChickWeight)
x <- sort(as.numeric(rownames(z)))
y <- sort(as.numeric(colnames(z)))
image(
	x = x, y = y, z = z, col = ice(100),
 xlab = 'Time', ylab = 'Chick'
)
}
\keyword{color}
