% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxy.R
\name{rain}
\alias{rain}
\title{Rain colour palette}
\arguments{
\item{n}{integer giving the number of colours (>= 1) to be produced;
currently it does not make sense to ask for more than 256}
}
\value{
Vector of R colours (\code{'#RRGGBB'} strings) of length \code{n}.
}
\description{
The rain colormap is sequential from light, dry colors to blue, wet colors, and could be used to plot amounts of rainfall.
}
\examples{
z <- xtabs(weight~Time+Chick, ChickWeight)
x <- sort(as.numeric(rownames(z)))
y <- sort(as.numeric(colnames(z)))
image(
	x = x, y = y, z = z, col = rain(100),
 xlab = 'Time', ylab = 'Chick'
)
}
\keyword{color}
