% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxy.R
\name{tempo}
\alias{tempo}
\title{Tempo colour palette}
\arguments{
\item{n}{integer giving the number of colours (>= 1) to be produced;
currently it does not make sense to ask for more than 256}
}
\value{
Vector of R colours (\code{'#RRGGBB'} strings) of length \code{n}.
}
\description{
The tempo colormap is sequential from whitish to dark teal and could be used to represent an increase in wave period values. This colormap is the negative half of the \code{\link{curl}} colormap.
}
\examples{
z <- xtabs(weight~Time+Chick, ChickWeight)
x <- sort(as.numeric(rownames(z)))
y <- sort(as.numeric(colnames(z)))
image(
	x = x, y = y, z = z, col = tempo(100),
 xlab = 'Time', ylab = 'Chick'
)
}
\keyword{color}
