% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxy.R
\name{topo}
\alias{topo}
\title{Topo colour palette}
\arguments{
\item{n}{integer giving the number of colours (>= 1) to be produced;
currently it does not make sense to ask for more than 256}
}
\value{
Vector of R colours (\code{'#RRGGBB'} strings) of length \code{n}.
}
\description{
The topo colormap has two distinct parts: one that is shades of blue and yellow to represent water depths (this is the \code{\link{deep}} colormap) and one that is shades of browns and greens to represent land elevation.
}
\examples{
z <- xtabs(weight~Time+Chick, ChickWeight)
x <- sort(as.numeric(rownames(z)))
y <- sort(as.numeric(colnames(z)))
image(
	x = x, y = y, z = z, col = topo(100),
 xlab = 'Time', ylab = 'Chick'
)
}
\keyword{color}
