\name{cmocean-package}
\alias{cmocean-package}
\alias{cmocean}
\docType{package}
\title{
\packageTitle{cmocean}
}
\description{
\packageDescription{cmocean}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cmocean}
\packageIndices{cmocean}

This package contains colormaps for commonly-used oceanographic
variables. Most of the colormaps started from matplotlib
colormaps, but have now been adjusted using the viscm tool to be
perceptually uniform.
}
\author{
\packageAuthor{cmocean}

Maintainer: \packageMaintainer{cmocean}
}
\references{
Thyng, K. M., Greene, C. A., Hetland, R. D., Zimmerle, H. M., &
DiMarco, S. F. (2016). True colors of oceanography. Oceanography,
29(3), 10.
}
\keyword{ package }
\keyword{ color }
