% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmocean.R
\docType{package}
\name{cmocean-package}
\alias{cmocean-package}
\title{cmocean: Beautiful Colourmaps for Oceanography}
\description{
Perceptually uniform palettes for commonly used
	variables in oceanography as functions taking an integer
	and producing character vectors	of colours.
}
\details{
This package contains colormaps for commonly-used oceanographic
variables. Most of the colormaps started from matplotlib
colormaps, but have now been adjusted using the viscm tool to be
perceptually uniform.
}
\references{
Thyng, K. M., Greene, C. A., Hetland, R. D., Zimmerle, H. M., &
DiMarco, S. F. (2016). True colors of oceanography. Oceanography,
29(3), 10.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://matplotlib.org/cmocean/}
  \item Report bugs at \url{https://github.com/aitap/cmocean}
}

}
\author{
\strong{Maintainer}: Ivan Krylov \email{krylov.r00t@gmail.com}

Authors:
\itemize{
  \item Kristen Thyng \email{kthyng@gmail.com}
}

Other contributors:
\itemize{
  \item Clark Richards \email{clark.richards@gmail.com} [contributor]
}

}
\keyword{color}
