% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxy.R
\docType{package}
\name{cmocean-package}
\alias{cmocean}
\alias{cmocean-package}
\title{cmocean: Beautiful Colormaps for Oceanography}
\description{
A (one paragraph) description of what
	the package does and why it may be useful.
}
\details{
This package contains colormaps for commonly-used oceanographic
variables. Most of the colormaps started from matplotlib
colormaps, but have now been adjusted using the viscm tool to be
perceptually uniform.
}
\references{
Thyng, K. M., Greene, C. A., Hetland, R. D., Zimmerle, H. M., &
DiMarco, S. F. (2016). True colors of oceanography. Oceanography,
29(3), 10.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://matplotlib.org/cmocean/}
  \item Report bugs at \url{https://github.com/aitap/cmocean}
}

}
\author{
\strong{Maintainer}: Ivan Krylov \email{krylov.r00t@gmail.com}

Authors:
\itemize{
  \item Kristen Thyng \email{kthyng@gmail.com}
}

}
\keyword{color}
