\name{cmocean-package}
\alias{cmocean-package}
\docType{package}
\title{
  \packageTitle{cmocean}
}
\description{
  \packageDescription{cmocean}
}
\references{
  Thyng, K. M., Greene, C. A., Hetland, R. D., Zimmerle, H. M., and
  DiMarco, S. F. (2016). True colors of oceanography. Oceanography,
  29(3), 10.
}
\seealso{
  \code{\link{cmocean}} for the classic-style pallette-generating
  function, \code{\link{scale_colour_cmocean}} for \pkg{ggplot2} helpers.
}
\keyword{color}
