\name{cmocean-package}
\alias{cmocean-package}
\docType{package}
\title{
\packageTitle{cmocean}
}
\description{
\packageDescription{cmocean}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cmocean}
\packageIndices{cmocean}
}
\author{
\packageAuthor{cmocean}

Maintainer: \packageMaintainer{cmocean}
}
\references{
Thyng, K. M., Greene, C. A., Hetland, R. D., Zimmerle, H. M., and
DiMarco, S. F. (2016). True colors of oceanography. Oceanography,
29(3), 10.
}
\keyword{color}
% vi:ts=2:et
